- removed compiler warnings https://github.com/openpeer/op/issues/153
- fixed date time routines to use specific date time and not generic "duration"
https://github.com/openpeer/op/issues/150

+++ openpeer/stack/IServiceSalt.h
@@ -83,7 +83,7 @@ namespace openpeer
                                  String *outErrorReason = NULL
                                  ) const = 0;
 
-      virtual UINT getTotalSignedSaltsAvailable() const = 0;
+      virtual size_t getTotalSignedSaltsAvailable() const = 0;
       virtual ElementPtr getNextSignedSalt() = 0;
 
       virtual void cancel() = 0;
